+++ src/pages/_app.tsx
@@ -1,17 +1,17 @@
 import { ChakraProvider } from '@chakra-ui/react'
 import { AppProps } from 'next/app'
 import { RecoilRoot } from 'recoil'
-import { Layout } from '@/components/layout'
 import { Meta } from '@/components/meta'
+import { DefaultLayout } from '@/components/layouts/DefaultLayout'
 
 export default function MyApp({ Component, pageProps }: AppProps) {
   return (
     <RecoilRoot>
       <ChakraProvider>
         <Meta />
-        <Layout>
+        <DefaultLayout>
           <Component {...pageProps} />
-        </Layout>
+        </DefaultLayout>
       </ChakraProvider>
     </RecoilRoot>
   )
